requirements.txt
websocket-client==1.5.1
pytest
pytest-html
